+++ src/tower.hpp
@@ -8,9 +8,9 @@
 class tower : public sf::RectangleShape{
 public:
 	const std::string name;
-	unsigned int damage;
-	unsigned int range;
-	unsigned int firerate;
+	double damage;
+	double range;
+	double firerate;
 	unsigned int value;
 
 	unsigned int upgrade = 1;
